+++ history.md
@@ -37,7 +37,7 @@ Semantic Versioning 2.0.0 is from version 0.1.6+
 
 ## version 0.5.12 _(Unreleased)_ - 2023-10-?? {#v0.5.12}
 
-- nothing yet
+- [x] (Changed) _Back-end_ Upgrade to .NET 6 - SDK 6.0.416 (Runtime: 6.0.24) (PR #1266)
 
 ## version 0.5.11 - 2023-10-13 {#v0.5.11}