+++ history.md
@@ -46,6 +46,7 @@ Semantic Versioning 2.0.0 is from version 0.1.6+
 - [x] (Fixed) _Back-end_ MySql timeouts bug (PR #1277, issue #1186)
 - [x] (Fixed) _Back-end_ Thumbnail CLI null reference bug  (PR #1277, issue #1176)
 - [x] (Fixed) _Back-end_ Fix code smells and add tests (PR #1278)
+- [x] (Fixed) _Back-end_ Fix builds for Windows (PR #1284)
 
 ## version 0.5.11 - 2023-10-13 {#v0.5.11}