+++ history.md
@@ -35,7 +35,11 @@ Semantic Versioning 2.0.0 is from version 0.1.6+
 
 ## List of versions
 
-## version 0.5.12 _(Unreleased)_ - 2023-10-?? {#v0.5.12}
+## version 0.5.13 _(Unreleased)_ - 2023-11-?? {#v0.5.13}
+
+- nothing yet
+
+## version 0.5.12 - 2023-11-17 {#v0.5.12}
 
 - [x] (Changed) _Back-end_ Upgrade to .NET 6 - SDK 6.0.416 (Runtime: 6.0.24) (PR #1266)
 - [x] (Changed) _Front-end_ improve accessibility and code smells (PR #1271, #1274)